+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - (Feature) Adds shortcuts `<ctrl-u>` and `<ctrl-d>` to scroll half a page up and down, respectively.
 - (Feature) Adds shortcuts `<ctrl-o>` and `<ctrl-i>` to navigate backwards and forwards, respectively, in tab history.
+- (Feature) Copies title of page to clipboard using `yt`.
 - (Chore) Dials back highlighting of selected links.
 
 ### 1.3.0